+++ src/langserver/languageInformation/toolsGetImagesRequest.ts
@@ -1,0 +1,7 @@
+export class ToolsGetImagesRequest {
+    type: string;
+
+    constructor(newType: string) {
+        this.type = newType;
+    }
+}