+++ src/langserver/languageInformation/toolsGetImagesResponse.ts
@@ -1,0 +1,5 @@
+import { ImageInformation } from "./imageInformation";
+
+export class ToolsGetImagesResponse {
+    images: ImageInformation[] | undefined;
+}