+++ CHANGELOG.md
@@ -3,13 +3,15 @@
 ## 3.0.6
  - Issue #88 - "Show Code Analyzer Rules" - include compiler rules (ALxxxx)
  - Issue #244 - Add ToolTip not working correct if ' is used
+ - Issue #246 - 'Show Action Images' does not work in a multi-root workspace
  - Issue #247 - Exclude specific object from codeActionsOnSave
  - Issue #250 - ToolTip is added when ToolTipML is already set
+ 
 
  Thank you
   - rvanbekkum for reporting issue #88
   - Duffy77 for reporting issue #244
-  - fvet for reporting issue #247
+  - fvet for reporting issue #246, #247
   - wbrakowski for reporting issue #250
   
 ## 3.0.5